[graco_sample] added explicit size_t cast to shut msvc14 warning

+++ samples/graco_sample_01_calc/Value.cpp
@@ -327,7 +327,7 @@ Value::relationalOperator (
 		return false;
 	}
 
-	ASSERT (opKind < countof (operatorTable));
+	ASSERT ((size_t) opKind < countof (operatorTable));
 	Operator* op = &operatorTable [opKind];
 
 	Type type = AXL_MAX (m_type, value.m_type);
